pyproject.toml
[project]
name = "replicode"
version = "0.1.0"
description = "A tool to score the reproducibility readiness of ML code repositories"
readme = "README_REPLICODE.md"
requires-python = ">=3.10"
dependencies = [
    "claude-agent-sdk==0.1.33",
    "datasets>=4.5.0",
    "google-genai>=1.62.0",
    "pygithub>=2.8.1",
    "rich>=10.0.0",
]
authors = [
    { name = "RepliCode Team" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

[project.scripts]
replicode = "replicode.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["replicode"]
